migrate to PEP-0420 to stop breaking sphinxcontrib namespace
This project currently breaks other packages in the sphinxcontrib
namespace if other packages are installed in different locations.

Namespaces may have multiple "portions" (regular packages accessible
under the namespace) which are allowed to be located in different
paths[1]. Installing this package breaks the sphinxcontrib namespace and
makes it impossible to find portions in other locations.

This can be highlighted by installing any sphinxcontrib package from
your host package manager (e.g. sphinxcontrib-jquery) and then creating
a venv with system packages:

```shell
python3 -m venv --system-site-packages venv/
. venv/bin/activate
```

Check that sphinxcontrib-jquery is available by running:

```shell
python3 -c "import sphinxcontrib.jquery"
```

Install sphinxcontrib-applehelp (or any other sphinxcontrib-namespaced
package not currently installed in your host distribution) with

```shell
python3 -m pip install sphinxcontrib-applehelp
```

Check it can be found by running

```shell
python3 -c "import sphinxcontrib.applehelp"
```

You can check where sphinxcontrib namespace is going to look for its
portions with

```shell
python3 -c 'import sphinxcontrib; print(sphinxcontrib.__path__)'
```

You'll see multiple paths there, including the host distribution's and
venv's.

After installing this package with

```shell
python3 -m pip install sphinxcontrib-svg2pdfconverter
```

Run the same commands. sphinxcontrib-jquery cannot be found anymore and
the sphinxcontrib.__path__ returns only one path now, the one where
svg2pdfconverter is installed.

Let's migrate this package to use PEP-0420, available since Python 3.3
(released Sep. 29, 2012), by making use of implicit namespacing (that
is, a directory without an __init__.py with subdirectories with an
__init__.py in it). Aside from fixing the sphinxcontrib namespace, no
intended change in behavior as the various sphinxcontrib python modules
(extensions in Sphinx terms) will still be available under the same
name.

The setup function is split from the class, the former being put in the
__init__.py of the module and the rest in the Python module named after
the converter.

[1] https://peps.python.org/pep-0420/#differences-between-namespace-packages-and-regular-packages
